logger handled successfully
