+++ assignment4/src/lib.rs
@@ -1,4 +1,5 @@
 #![allow(unused)]
 mod hello_count;
-mod pelindrom;
+mod palindrom;
 mod rotation;
+mod test;
